refactor(export): extract sanitizeFilename + createZipArchive + getProgressMessage
`exportService.ts` shrinks from 1824 → 1651 LoC (-173) by hoisting three
self-contained utilities into `export/exportFileOperations.ts`:

- `sanitizeFilename` — strip filesystem-unsafe chars + Windows reserved
  device names. Pure function. Now also reusable outside ExportService.
- `getProgressMessage` — build human-readable WebSocket progress strings.
  Pure function with no dependencies. Type aliases `ExportProgressStage`
  and `ExportProgressDetail` are exported for reuse.
- `createZipArchive` — async filesystem op that writes a zip from a
  staging directory to `process.env.EXPORT_DIR || './exports'`. Uses
  `archiver` + `fs.promises`; idempotent cleanup guards against
  memory leaks across many exports. Class call site dropped the
  `this.sanitizeFilename(...)` indirection — the function imports
  `sanitizeFilename` directly now.

The class shell still owns job lifecycle (`exportJobs` map, retry, WS
notifications, cleanup interval, progress tracking via
`updateJobProgress`/`isJobCancelled`/`sendToUser`).

Verification:
- Backend `npx tsc --noEmit` passes
- `exportService.test.ts` + `exportController.test.ts`: 41/41 tests pass

Cumulative `exportService.ts` reduction across this session:
2480 → 1824 (PR #118 docs) → 1651 (this PR) = -829 LoC (-33%)
